Move tests
- rename corpora-api.yml to portal-api.yml
