+++ tests/unit/backend/common/entities/datasets/test_create.py
@@ -4,7 +4,7 @@ from backend.common.corpora_orm import (
     ValidationStatus,
 )
 from backend.common.entities import Dataset
-from tests.unit.backend.corpora.common.entities.datasets import TestDataset
+from tests.unit.backend.base_test_dataset import TestDataset
 from tests.unit.backend.utils import BogusDatasetParams
