+++ tests/unit/backend/common/entities/datasets/test_get.py
@@ -1,6 +1,6 @@
 from backend.common.corpora_orm import DbDatasetArtifact, DbDatasetProcessingStatus
 from backend.common.entities import Dataset
-from tests.unit.backend.corpora.common.entities.datasets import TestDataset
+from tests.unit.backend.base_test_dataset import TestDataset
 
 
 class TestGetDataset(TestDataset):
