+++ tests/unit/backend/common/entities/datasets/test_list.py
@@ -1,5 +1,5 @@
 from backend.common.entities import Dataset
-from tests.unit.backend.corpora.common.entities.datasets import TestDataset
+from tests.unit.backend.base_test_dataset import TestDataset
 from tests.unit.backend.utils import BogusDatasetParams
