+++ tests/unit/backend/common/entities/datasets/test_reprocess.py
@@ -1,4 +1,4 @@
-from tests.unit.backend.corpora.common.entities.datasets import TestDataset
+from tests.unit.backend.base_test_dataset import TestDataset
 
 
 class TestReprocessDataset(TestDataset):
