+++ tests/unit/backend/common/utils/dl_sources/test_url.py
@@ -8,7 +8,7 @@ import requests
 from moto import mock_s3
 
 from backend.common.utils.dl_sources.url import from_url, RegisteredSources, S3URL, DropBoxURL, S3URI
-from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
+from tests.unit.backend.mock_aws_test_case import CorporaTestCaseUsingMockAWS
 
 
 class TestRegisteredSources(unittest.TestCase):
