+++ tests/unit/backend/dataset_processing/test_downloader.py
@@ -12,7 +12,7 @@ from backend.common.entities import Dataset
 from backend.common.utils.math_utils import MB
 from backend.dataset_processing import download
 from backend.dataset_processing.exceptions import ProcessingFailed
-from tests.unit.backend.fixtures.data_portal_test_case import DataPortalTestCase
+from tests.unit.backend.data_portal_test_case import DataPortalTestCase
 
 
 def start_server(path, port):
