+++ tests/unit/backend/portal/api_server/collection_id/test_create_genesets.py
@@ -1,7 +1,7 @@
 import json
 from backend.common.corpora_orm import CollectionVisibility, generate_id
 from backend.common.entities.geneset import Geneset
-from tests.unit.backend.corpora.api_server.base_api_test import BaseAuthAPITest, get_cxguser_token
+from tests.unit.backend.portal.api_server.base_api_test import BaseAuthAPITest, get_cxguser_token
 
 
 class TestGenesetCreation(BaseAuthAPITest):
