+++ tests/unit/backend/portal/api_server/curator/collection/datasets/test_status.py
@@ -1,5 +1,5 @@
 from backend.common.corpora_orm import CollectionVisibility, ProcessingStatus
-from tests.unit.backend.corpora.api_server.base_api_test import BaseAuthAPITest
+from tests.unit.backend.portal.api_server.base_api_test import BaseAuthAPITest
 
 
 class TestDatasetStatus(BaseAuthAPITest):
