+++ tests/unit/backend/portal/api_server/curator/collection/test_assets.py
@@ -1,9 +1,9 @@
 from unittest.mock import patch, Mock
 
 from backend.common.corpora_orm import CollectionVisibility, DatasetArtifactFileType
-from tests.unit.backend.corpora.api_server.base_api_test import BaseAuthAPITest
+from tests.unit.backend.portal.api_server.base_api_test import BaseAuthAPITest
 from tests.unit.backend.fixtures.config import fake_s3_file
-from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
+from tests.unit.backend.mock_aws_test_case import CorporaTestCaseUsingMockAWS
 
 
 # TODO test with dataset_id = test_curator_tag
