+++ tests/unit/backend/portal/api_server/curator/collection/test_collection_delete.py
@@ -1,7 +1,7 @@
 import unittest
 
 from backend.common.corpora_orm import CollectionVisibility
-from tests.unit.backend.corpora.api_server.base_api_test import BaseAuthAPITest
+from tests.unit.backend.portal.api_server.base_api_test import BaseAuthAPITest
 
 
 class TestDeleteCollection(BaseAuthAPITest):
