+++ tests/unit/backend/portal/api_server/curator/collection/test_curation_revision.py
@@ -1,5 +1,5 @@
 from backend.common.corpora_orm import CollectionVisibility
-from tests.unit.backend.corpora.api_server.base_api_test import BaseAuthAPITest
+from tests.unit.backend.portal.api_server.base_api_test import BaseAuthAPITest
 
 
 class TestPostRevision(BaseAuthAPITest):
