+++ tests/unit/backend/portal/api_server/mock_auth.py
@@ -8,7 +8,7 @@ from multiprocessing import Process
 
 # seconds until the token expires
 from jose import jwt
-from tests.unit.backend.corpora.api_server.config import TOKEN_EXPIRES
+from tests.unit.backend.portal.api_server.config import TOKEN_EXPIRES
 
 
 # A mocked out oauth server, which serves all the endpoints needed by the oauth type.
