+++ tests/unit/backend/portal/api_server/test_v1_auth_keys.py
@@ -4,7 +4,7 @@ from jose import jws
 from unittest.mock import patch, DEFAULT
 
 from backend.common.auth0_manager import auth0_management_session
-from tests.unit.backend.corpora.api_server.base_api_test import BaseAuthAPITest, get_cxguser_token
+from tests.unit.backend.portal.api_server.base_api_test import BaseAuthAPITest, get_cxguser_token
 
 
 class TestKeys(BaseAuthAPITest):
