+++ tests/unit/backend/portal/api_server/test_v1_genesets.py
@@ -2,8 +2,8 @@ import json
 
 from backend.common.corpora_orm import CollectionVisibility
 from backend.common.entities.geneset import GenesetDatasetLink
-from tests.unit.backend.corpora.api_server.base_api_test import BaseAuthAPITest, get_cxguser_token
-from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
+from tests.unit.backend.portal.api_server.base_api_test import BaseAuthAPITest, get_cxguser_token
+from tests.unit.backend.mock_aws_test_case import CorporaTestCaseUsingMockAWS
 
 
 class TestGenesets(BaseAuthAPITest, CorporaTestCaseUsingMockAWS):
