+++ tests/unit/backend/wmg/data/test_extract.py
@@ -3,8 +3,8 @@ import unittest
 import backend.corpus_asset_pipelines.integrated_corpus.extract
 from backend.common.corpora_orm import DatasetArtifactFileType
 from backend.wmg.data.constants import INCLUDED_ASSAYS
-from tests.unit.backend.fixtures.generate_data_mixin import GenerateDataMixin
-from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
+from tests.unit.backend.generate_data_mixin import GenerateDataMixin
+from tests.unit.backend.mock_aws_test_case import CorporaTestCaseUsingMockAWS
 
 
 class TestExtract(CorporaTestCaseUsingMockAWS, GenerateDataMixin):
